+++ polars_bio/__init__.py
@@ -51,6 +51,7 @@ register_gff = data_processing.register_gff
 register_gtf = data_processing.register_gtf
 register_vcf = data_processing.register_vcf
 register_bcf = data_processing.register_bcf
+register_bgen = data_processing.register_bgen
 register_vcf_zarr = data_processing.register_vcf_zarr
 register_fastq = data_processing.register_fastq
 register_fasta = data_processing.register_fasta
@@ -67,6 +68,7 @@ sql = data_processing.sql
 
 describe_vcf = data_input.describe_vcf
 describe_bcf = data_input.describe_bcf
+describe_bgen = data_input.describe_bgen
 describe_vcf_zarr = data_input.describe_vcf_zarr
 describe_bam = data_input.describe_bam
 describe_sam = data_input.describe_sam
@@ -80,6 +82,7 @@ read_gff = data_input.read_gff
 read_gtf = data_input.read_gtf
 read_table = data_input.read_table
 read_bcf = data_input.read_bcf
+read_bgen = data_input.read_bgen
 read_vcf = data_input.read_vcf
 read_vcf_zarr = data_input.read_vcf_zarr
 read_fastq = data_input.read_fastq
@@ -101,6 +104,7 @@ scan_gtf = data_input.scan_gtf
 scan_pairs = data_input.scan_pairs
 scan_table = data_input.scan_table
 scan_bcf = data_input.scan_bcf
+scan_bgen = data_input.scan_bgen
 scan_vcf = data_input.scan_vcf
 scan_vcf_zarr = data_input.scan_vcf_zarr
 write_vcf = data_input.write_vcf
@@ -162,6 +166,7 @@ __all__ = [
     "print_metadata_summary",
     # I/O functions
     "describe_bcf",
+    "describe_bgen",
     "describe_vcf",
     "describe_vcf_zarr",
     "describe_bam",
@@ -181,6 +186,7 @@ __all__ = [
     "read_gtf",
     "read_table",
     "read_bcf",
+    "read_bgen",
     "read_vcf",
     "read_vcf_zarr",
     "scan_bam",
@@ -196,6 +202,7 @@ __all__ = [
     "scan_pairs",
     "scan_table",
     "scan_bcf",
+    "scan_bgen",
     "scan_vcf",
     "scan_vcf_zarr",
     "write_vcf",
@@ -213,6 +220,7 @@ __all__ = [
     "register_gff",
     "register_gtf",
     "register_bcf",
+    "register_bgen",
     "register_vcf",
     "register_vcf_zarr",
     "register_fastq",