+++ src/lib.rs
@@ -43,11 +43,11 @@ use crate::context::PyBioSessionContext;
 use crate::operation::do_range_operation;
 use crate::option::{
     pyobject_storage_options_to_object_storage_options, BamReadOptions, BamWriteOptions,
-    BedReadOptions, BigBedReadOptions, BigWigReadOptions, BioTable, CramReadOptions,
-    CramWriteOptions, FastaReadOptions, FastaWriteOptions, FastqReadOptions, FastqWriteOptions,
-    FilterOp, GffReadOptions, GtfReadOptions, InputFormat, OutputFormat, OverlapOutputMode,
-    PairsReadOptions, PileupOptions, PyObjectStorageOptions, RangeOp, RangeOptions, ReadOptions,
-    VcfReadOptions, VcfWriteOptions, VcfZarrReadOptions, WriteOptions,
+    BedReadOptions, BgenReadOptions, BigBedReadOptions, BigWigReadOptions, BioTable,
+    CramReadOptions, CramWriteOptions, FastaReadOptions, FastaWriteOptions, FastqReadOptions,
+    FastqWriteOptions, FilterOp, GffReadOptions, GtfReadOptions, InputFormat, OutputFormat,
+    OverlapOutputMode, PairsReadOptions, PileupOptions, PyObjectStorageOptions, RangeOp,
+    RangeOptions, ReadOptions, VcfReadOptions, VcfWriteOptions, VcfZarrReadOptions, WriteOptions,
 };
 use crate::scan::{
     maybe_register_table, register_frame, register_frame_from_arrow_stream,
@@ -955,6 +955,7 @@ fn polars_bio(_py: Python, m: &Bound<PyModule>) -> PyResult<()> {
     m.add_class::<FastaReadOptions>()?;
     m.add_class::<FastaWriteOptions>()?;
     m.add_class::<PairsReadOptions>()?;
+    m.add_class::<BgenReadOptions>()?;
     m.add_class::<PileupOptions>()?;
     m.add_class::<PyObjectStorageOptions>()?;
     Ok(())